Small fix on code
on line 247:
if( ! (apply(dep.data, 2, is.numeric)) || apply(indep.data, 2, is.numeric)) ) 
was giving an error, since the logical argument was returning a logical vector that could not be interpreted by if() in CHECK but was interpreted as a all(logical_vector) on the console. So I added the all() around each apply.